Refactor(game): Split game.js into modules
Refactors the large game.js file into smaller, more manageable modules based on functionality:
- definitions.js: Contains static definitions for buildings, staff, and upgrades.
- state.js: Manages the core gameState object and save/load/reset logic.
- ui.js: Handles all DOM manipulation, rendering, and UI updates.
- gameLogic.js: Contains the main game loop, click handling, purchasing logic, and level/unlock mechanics.
- game.js: Acts as the main entry point, importing modules and initializing the game.

Updates index.html to load game.js as a module.

Addresses issue #5.
