+++ index.html
@@ -98,6 +98,6 @@
         </footer>
     </div>
 
-    <script src="game.js"></script>
+    <script type="module" src="game.js"></script>
 </body>
 </html>
